+++ src/receiverFMQ.cxx
@@ -33,6 +33,7 @@
 #include "SubTimeframe.h"
 
 #include "RdhUtils.h"
+#include "DataBlock.h"
 
 // definition of a global for logging
 using namespace AliceO2::InfoLogger;
@@ -258,7 +259,7 @@ int main(int argc, const char **argv) {
   std::string cfgDecodingMode = "none";
   cfg.getOptionalValue<std::string>(cfgEntryPoint + ".decodingMode",
                                     cfgDecodingMode);
-  enum decodingMode { none = 0, stfHbf = 1, stfSuperpage = 2 };
+  enum decodingMode { none = 0, stfHbf = 1, stfSuperpage = 2, stfDatablock = 3 };
   decodingMode mode = decodingMode::none;
   if (cfgDecodingMode == "none") {
     mode = decodingMode::none;
@@ -266,6 +267,8 @@ int main(int argc, const char **argv) {
     mode = decodingMode::stfHbf;
   } else if (cfgDecodingMode == "stfSuperpage") {
     mode = decodingMode::stfSuperpage;
+  } else if (cfgDecodingMode == "stfDatablock") {
+    mode = decodingMode::stfDatablock;
   } else {
     theLog.log(InfoLogger::Severity::Error, "Wrong decoding mode set : %s",
                cfgDecodingMode.c_str());
@@ -314,7 +317,7 @@ int main(int argc, const char **argv) {
   double copyRatio=0;
   unsigned long long copyRatioCount=0;
 
-  if ( (mode == decodingMode::stfHbf) || (mode == decodingMode::stfSuperpage)){
+  if ( (mode == decodingMode::stfHbf) || (mode == decodingMode::stfSuperpage) || (mode == decodingMode::stfDatablock)){
     isMultiPart = true;
   }
 
@@ -348,6 +351,7 @@ int main(int argc, const char **argv) {
           int i = 0;
           bool dumpNext = false;
           SubTimeframe *stf = nullptr;
+          int numberOfHBF = nPart - 1;
           for (auto const &mm : msgParts) {
             
             if (i == 0) {
@@ -359,14 +363,6 @@ int main(int argc, const char **argv) {
                 break;
               }
               stf = (SubTimeframe *)mm->GetData();
-              if (((int)stf->numberOfHBF != nPart - 1) &&  (stf->numberOfHBF != 0)) {
-                theLog.log(
-                    InfoLogger::Severity::Error,
-                    "Mismatch stf->numberOfHBF %d != %d message parts - 1\n",
-                    stf->numberOfHBF, nPart - 1);
-                break;
-              }
-              // number of message parts matches number of HBFs in header ?
               if (cfgDumpTF) {
                 if ((stf->timeframeId == 1) ||
                     (stf->timeframeId % cfgDumpTF == 0)) {
@@ -374,7 +370,7 @@ int main(int argc, const char **argv) {
                 }
               }
             } else {	  
-              if (stf->numberOfHBF != 0)  {
+              if ((numberOfHBF != 0) && (stf->isRdhFormat))  {
 	      // then we have 1 part per HBF
               size_t dataSize = mm->GetSize();
               void *data = mm->GetData();
@@ -391,7 +387,7 @@ int main(int argc, const char **argv) {
                   printf("Receiving TF %d CRU %d.%d link %d : %d HBf\n",
                          (int)stf->timeframeId, (int)h.getCruId(),
 			 (int)h.getEndPointId(),
-                         (int)stf->linkId, (int)stf->numberOfHBF);
+                         (int)stf->linkId, numberOfHBF);
                   dumpNext = false;
                 }
 
@@ -456,6 +452,22 @@ int main(int argc, const char **argv) {
 	      i++;
 	    }
           }
+	} else if (mode == decodingMode::stfDatablock) {
+	  nMsgParts += msgParts.size();
+	  //printf("parts=%d\n",msgParts.size());
+	  if (msgParts.size()!=2) {
+	    theLog.log(InfoLogger::Severity::Error,
+                       "%d parts in message, should be 2", (int)msgParts.size());
+	  } else {
+	    int sz = msgParts[0]->GetSize();
+	    if (sz != sizeof(DataBlockHeaderBase)) {
+              theLog.log(InfoLogger::Severity::Error,
+                       "part[0] size = %d, should be %d",sz , (int)sizeof(DataBlock));
+	    } else {
+	      DataBlockHeaderBase *dbhb = (DataBlockHeaderBase *)msgParts[0]->GetData();
+	      // printf("rx datablock size: header %d ?= msgpart %d\n",(int)dbhb->dataSize,(int)msgParts[1]->GetSize());
+	    }
+	  }
 	}
       }
     } else {